fix(ci): restore monitor runtime hotspot budgets
