+++ native/rust/crates/ripdpi-monitor-engine/src/engine/runtime/cancellation.rs
@@ -9,10 +9,7 @@ use super::plan::ExecutionPlan;
 use super::state::ExecutionRuntime;
 
 mod progress;
-
-pub(in crate::engine) fn cancelled_run_summary(has_partial_results: bool) -> &'static str {
-    if has_partial_results { "Scan completed with partial results" } else { "Scan cancelled" }
-}
+pub(in crate::engine) use progress::cancelled_run_summary;
 
 pub(in crate::engine) fn publish_cancelled_run(
     plan: &ExecutionPlan,
